Refactor Categories Provider
