+++ app/gui/src/dashboard/layouts/Drive/Categories/index.ts
@@ -1,3 +1,4 @@
 export * from './categoriesHooks'
+export * from './CategoriesProvider'
 export * from './Category'
 export * from './transferBetweenCategoriesHooks'
